Change iptable rule management logic.
The current way the formula works does not play well with other
apps that change iptables rules on their own.

This change changes the logic and inserts rules in the current
ruleset and then saves the current state as is, instead of generating
the full ruleset in a template and leave additional changes out.